Hook implemented
The hook works in pairs. You throw one, and it sets up your next item to be another hook. Hook1 is the compliment to Hook2.

If a hook is out to long it will destroy itself and the compliment. If the hooks are too far away it will also self destruct