A Pick & Map Generation
Main feature of the day, a swingable pick. Had to do a little tweaking in a few places, figure out how things communicated with eachother.

Having the rigid body on the pick in the same location as the Animator, so I know have a member variable in the collider and in the controller. I had to do it this way because unlike the kunai or the hook, the pick is controlled by the animator, so Having a character controller would have messed with that.

 I'm not entirely sure why, but now the animator on all the blocks is causing the game to lag. I mean it makes sense considering their are thoousands of blocks that are just doing nothing. But the weird thing is that despite the culling and custom culling attempts, it still runs slow. My plan is to display blocks in a way that does not require the animator, and try to stress test the # of animators on the screen and see if I can get the culling to work properly, cause on camer there should only be a handful of blocks.